requirements.txt
openai
python-dotenv
numpy
sentence-transformers
